isolate core module and extend core in index.js
