+++ src/httpStore.js
@@ -1,38 +1,32 @@
 import { KeyError } from './errors.js';
 
-function joinUrlParts(...args) {
-  return args.map((part, i) => {
-    if (i === 0) {
-      return part.trim().replace(/[\/]*$/g, '');
-    } else {
-      return part.trim().replace(/(^[\/]*|[\/]*$)/g, '');
-    }
-  }).filter(x=>x.length).join('/');
-}
+// Default HTTPStore is read-only
 
 export default class HTTPStore {
   constructor(url) {
-    this.url = url;
+    this.url =  url.endsWith('/') ? url.slice(0, -1) : url;
+  }
+
+  _path(key) {
+    return `${this.url}/${key.startsWith('/') ? key.slice(1) : key}`;
   }
   
   async getItem(key) {
-    const url = joinUrlParts(this.url, key);
-    const value = await fetch(url);
-
+    const path = this._path(key);
+    const value = await fetch(path);
     if (value.status === 404) {
       // Item is not found
       throw new KeyError(key);
     } else if (value.status !== 200) {
-      throw new HTTPError(String(value.status));
+      throw new Error('HTTPError:', String(value.status));
     }
     // only decode if 200
-    return value.arrayBuffer(); // Browser
+    return value.arrayBuffer();
   }
   
   async containsItem(key) {
-    const url = joinUrlParts(this.url, key);
-    // Just check headers if HEAD method supported
-    const value = await fetch(url);
+    const path = this._path(key);
+    const value = await fetch(path);
     return value.status === 200;
   }
 
@@ -40,7 +34,7 @@ export default class HTTPStore {
     throw new Error('Method not impemented.');
   }
 
-  async setItem(item, value) {
+  setItem(item, value) {
     throw new Error('Method not implemented.')
   }
