--- src/indexing.js
@@ -1,21 +1,0 @@
-export * from './index.js';
-import { ZarrArray } from './index.js';
-
-import { _BasicIndexer, _get_selection, slice } from '../node_modules/zarrita/src/indexing.js';
-export { slice };
-
-// mutate prototype and add indexing
-
-Object.defineProperties(ZarrArray.prototype, {
-  getRaw: {
-    value: function(selection) {
-      return this.getBasicSelection(selection);
-    },
-  },
-  getBasicSelection: {
-    value: function(selection) {
-      const indexer = new _BasicIndexer({ selection, ...this });
-      return _get_selection.call(this, indexer);
-    }
-  }
-});
